Move type checker to own file
